add scroll function after show/hide a phrase

+++ pages/phrases.php
@@ -20,10 +20,25 @@ if(empty($phrases)) {
 }
 
 if ($_SERVER["REQUEST_METHOD"] === "POST") {
-    $id = htmlspecialchars($_POST["delete_id"]);
-    Phrases::delete($connection, $id);
-    header("Location: {$_SERVER['PHP_SELF']}");
-    exit();
+    if(isset($_POST['delete'])) {
+        $id = htmlspecialchars($_POST["delete_id"]);
+        Phrases::delete($connection, $id);
+        header("Location: {$_SERVER['PHP_SELF']}");
+        exit();
+    }
+    elseif(isset($_POST['hide'])) {
+        $id = htmlspecialchars($_POST["id"]);
+        Phrases::showToggle($connection, $id, 'hide');
+        header("Location: {$_SERVER['PHP_SELF']}?scroll=phrase-$id");
+        exit();
+    }
+
+    elseif(isset($_POST['show'])) {
+        $id = htmlspecialchars($_POST["id"]);
+        Phrases::showToggle($connection, $id, 'show');
+        header("Location: {$_SERVER['PHP_SELF']}?scroll=phrase-$id");
+        exit();
+    }
 }
 
 $filtered_phrases = [];
@@ -58,16 +73,35 @@ foreach ($phrases as $phrase) {
 
             <div class="all-phrases">
                 <?php foreach($filtered_phrases as $phrase): ?>
-                    <div class="one-phrase">
+                    <div class="one-phrase <?php echo $phrase['status'] === 'hide' ? 'hide' : ''; ?>" id="phrase-<?= $phrase["id_phrase"] ?>">
                         <div class="phrases">
                             <p><?= $phrase["slovak"] ?></p>
                             <p><?= $phrase["english"] ?></p>
                         </div>
                         <div class="btns">
+                            <?php if($phrase["status"] === 'show'): ?>
+                                <form method="post" action="?scroll=phrase-<?= $phrase["id_phrase"] ?>">
+                                    <input type="hidden" name="id" value="<?= $phrase["id_phrase"] ?>">
+                                    
+                                    <button type="submit" class="hide" name="hide">
+                                        <i class="fa-solid fa-eye-slash"></i>
+                                    </button>
+                                </form>
+                            <?php elseif($phrase["status"] === 'hide'): ?>
+                                <form method="post" action="?scroll=phrase-<?= $phrase["id_phrase"] ?>">
+                                    <input type="hidden" name="id" value="<?= $phrase["id_phrase"] ?>">
+                                    
+                                    <button type="submit" class="show" name="show">
+                                        <i class="fa-solid fa-eye"></i>
+                                    </button>
+                                </form>
+                            <?php endif; ?>
+                            
+
                             <form method="post">
                                 <input type="hidden" name="delete_id" value="<?= $phrase["id_phrase"] ?>">
                                 
-                                <button type="submit" class="delete">
+                                <button type="submit" class="delete" name="delete">
                                     <i class="fa-solid fa-trash"></i>
                                 </button>
                             </form>
@@ -85,5 +119,6 @@ foreach ($phrases as $phrase) {
     </div>
 
     <script src="../js/filter.js"></script>
+    <script src="../js/scroll-to-phrase.js"></script>
 </body>
 </html>
